requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
streamlit==1.30.0
pandas==2.2.2
plotly==5.18.0